refactor: modularize key_pair standards and forward subsystems into dedicated directories
- Split monolithic ssh_key_pair.dart into key_pair/openssh_key_pair.dart, key_pair/pkcs1_rsa_key_pair.dart, and key_pair/sec1_ec_key_pair.dart
- Move dynamic_forward and ssh_forward implementations into forward/ directory
- Update dartssh2.dart and internal imports to maintain 100% backwards compatibility
