+++ lib/dartssh2.dart
@@ -2,7 +2,6 @@ export 'src/ssh_algorithm.dart' show SSHAlgorithms;
 export 'src/ssh_agent.dart';
 export 'src/ssh_client.dart';
 export 'src/ssh_errors.dart';
-export 'src/ssh_forward.dart';
 export 'src/ssh_hostkey.dart';
 export 'src/ssh_identity.dart';
 export 'src/ssh_key_pair.dart';
@@ -12,12 +11,18 @@ export 'src/ssh_signal.dart';
 export 'src/ssh_transport.dart';
 export 'src/ssh_userauth.dart';
 
+export 'src/forward/ssh_forward.dart';
+export 'src/forward/dynamic_forward.dart';
 export 'src/socket/ssh_socket.dart';
 
 export 'src/algorithm/ssh_cipher_type.dart';
 export 'src/algorithm/ssh_hostkey_type.dart';
 export 'src/algorithm/ssh_kex_type.dart';
 export 'src/algorithm/ssh_mac_type.dart';
+
+export 'src/key_pair/openssh_key_pair.dart';
+export 'src/key_pair/pkcs1_rsa_key_pair.dart';
+export 'src/key_pair/sec1_ec_key_pair.dart';
 
 export 'src/sftp/sftp_client.dart';
 export 'src/sftp/sftp_errors.dart';
