+++ lib/src/forward/dynamic_forward_io.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:dartssh2/src/ssh_forward.dart';
+import 'package:dartssh2/src/forward/ssh_forward.dart';
 
 typedef SSHDynamicDial = Future<SSHForwardChannel> Function(
   String host,
@@ -389,7 +389,6 @@ class _ByteBuffer {
 class _TargetAddress {
   final String host;
   final int port;
-
   const _TargetAddress(this.host, this.port);
 }
