+++ test/src/socket/dynamic_forward_io_test.dart
@@ -3,10 +3,10 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:dartssh2/src/dynamic_forward.dart';
+import 'package:dartssh2/src/forward/dynamic_forward.dart';
 import 'package:dartssh2/src/message/msg_channel.dart';
 import 'package:dartssh2/src/ssh_channel.dart';
-import 'package:dartssh2/src/ssh_forward.dart';
+import 'package:dartssh2/src/forward/ssh_forward.dart';
 import 'package:test/test.dart';
 
 void main() {
